client, server, templates: Clean up YouTube embed function and implement BitChute embed
